Moved Strm file creation to Utils
